pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "orbital-radar"
version = "0.0.2"
authors = [
  { name = "Nils Risse", email = "n.risse@uni-koeln.de" },
  { name = "Lukas Pfitzenmaier", email = "l.pfitzenmaier@uni-koeln.de" },
  { name = "Pavlos Kollias" },
  { name = "Bernat Puigdomenech Treserras" },
]
description = "This repository contains the sub-orbital to orbital transformation tool for radar observations, enabling the conversion of radar data from ground or aircraft perspectives to the satellite viewpoint."
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "dask >= 2023.1.0",
  "ipykernel >= 6.25.0",
  "matplotlib >= 3.7.0",
  "netCDF4 >= 1.6.2",
  "numpy >= 1.25.0",
  "pandas >= 2.0.0",
  "pytest >= 7.4.0",
  "python-dotenv == 1.0.0",
  "scipy >= 1.10.0",
  "toml >= 0.10.2",
  "xarray >= 2023.01.0",
]

[project.optional-dependencies]
dev = ["numpydoc", "sphinx", "sphinx_rtd_theme", "tox", "black"]

[project.scripts]
orbital-radar = "orbital_radar.__main__:main"

[project.urls]
documentation = "INSERT_URL"

[tool.black]
line-length = 79
